backend/pyproject.toml
[project]
name = "leave-application-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[google-genai]>=0.3.25",
    "langchain-community>=0.3.24",
    "langgraph>=0.4.7",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
]
